refactor: Modulize UI logic into distinct JavaScript functions
Breaks down monolithic JavaScript logic into more organized and reusable
functions for improved clarity and maintainability.
