--- index.css
@@ -1,260 +1,0 @@
-@import url(global.css);
-
-.container-main {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-}
-/* title bar */
-.title-bar {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-  padding: var(--padding-small);
-}
-/* New wrapper for the right-aligned controls */
-.title-controls {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: var(--padding-medium); /* space between search and theme button */
-}
-/* logo */
-.title-logo {
-  font-size: var(--font-size-large);
-  font-weight: bold;
-  padding: var(--padding-small);
-}
-/* light dark theme button */
-.theme-button {
-  height: 50px;
-  width: 50px;
-}
-/* search bar */
-.title-search {
-  display: flex;
-  position: relative;
-  justify-content: space-between;
-  align-items: center;
-  height: 40px;
-  width: 280px;
-  padding: var(--padding-small);
-  border-radius: 28px;
-  background-color: var(--color-text-default);
-}
-/* search-bar input */
-.search-text {
-  border: none;
-  background: none;
-  outline: none;
-  float: left;
-  padding-left: 15px;
-  color: var(--color-bg-secondary);
-  font-size: var(--font-size-small);
-  font-family: var(--font-family-primary), serif;
-  transition: ease-in 0.4s;
-  line-height: 40px;
-  width: 250px;
-}
-/* search-bar button */
-#search-button {
-  color: var(--color-text-default);
-  float: right;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background: var(--color-bg-secondary);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-}
-/* --- landing section body --- */
-.body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  padding: 150px;
-}
-/* landing section main text content */
-.heading-main,
-.paragraph-main {
-  text-align: center;
-}
-/* create an account button */
-.start-button {
-  font-size: var(--font-size-medium);
-  height: 50px;
-  width: 180px;
-  border-radius: 28px !important;
-}
-.start-button:hover {
-  color: var(--color-success);
-}
-/* --- create account section --- */
-.create-account-section {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-/* create account main headding */
-.create-account-section p:nth-child(1) {
-  font-size: var(--font-size-large);
-  text-align: center;
-  margin: 0;
-}
-
-/* create account catchy sentence */
-.create-account-section p:nth-child(2) {
-  font-size: var(--font-size-small);
-  text-align: center;
-  margin: 0;
-}
-
-/* create account form */
-.create-account-form {
-  display: flex;
-  flex-direction: column;
-  margin: var(--margin-small);
-  width: 300px;
-}
-
-/* create account form label */
-.create-account-section label {
-  margin: var(--margin-small);
-  font-size: var(--font-size-small);
-  padding-left: 15px;
-}
-
-/* create account form input */
-.create-account-form input:not([type="radio"]) {
-  border: none;
-  padding: var(--padding-small);
-  border-radius: 28px;
-  background-color: var(--color-text-default);
-  padding-left: 15px;
-  color: var(--color-bg-secondary);
-  font-size: var(--font-size-small);
-  font-family: var(--font-family-primary), serif;
-  transition: ease-in 0.4s;
-  line-height: 30px;
-  width: 250px;
-  margin: var(--margin-small);
-}
-
-/* create account form input date */
-.create-account-form input[type="date"]::-webkit-calendar-picker-indicator {
-  background: transparent;
-  bottom: 0;
-  color: transparent;
-  cursor: pointer;
-  height: auto;
-  left: 0;
-  position: absolute;
-  right: 0;
-  top: 0;
-  width: auto;
-}
-
-.create-account-form input[type="date"] {
-  text-align: left !important;
-}
-
-/* Gender fieldset alignment */
-.gender-group {
-  border: none;
-  margin: var(--margin-small);
-  width: 250px;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: var(--padding-small);
-
-  margin-left: calc(var(--margin-small) + 15px);
-}
-
-.gender-group-row {
-  display: flex;
-  gap: var(--padding-xlarge);
-}
-
-.gender-group legend {
-  font-size: var(--font-size-small);
-  color: var(--color-text-default);
-  padding: 0;
-  margin: var(--margin-small);
-}
-
-/* Gender option containers */
-.gender-option {
-  display: flex;
-  align-items: center;
-  gap: var(--padding-small);
-}
-
-/* Radio button style */
-.gender-option input[type="radio"] {
-  width: 18px;
-  height: 18px;
-  border-radius: 12px;
-  cursor: pointer;
-  appearance: none;
-  background-color: var(--color-bg-default);
-  border: 2px solid var(--color-border-default);
-  transition: all 0.3s ease;
-  margin: 0; /* Remove default margin */
-}
-
-.gender-option input[type="radio"]:checked {
-  background-color: var(--color-accent);
-  border-color: var(--color-accent);
-}
-
-.gender-option input[type="radio"]:focus {
-  outline: none;
-  box-shadow: 0 0 0 3px rgba(0, 123, 255, 0.2);
-}
-
-/* Label styling */
-.gender-option label {
-  font-size: var(--font-size-small);
-  color: var(--color-text-default);
-  cursor: pointer;
-  margin: 0; /* Remove default margin */
-  padding: 0; /* Remove padding to align properly */
-  user-select: none;
-}
-
-.gender-option label:hover {
-  color: var(--color-accent);
-}
-
-/* create account form button */
-.form-buttons {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.back-button,
-.create-button {
-  font-size: var(--font-size-small);
-  height: 32px !important;
-  width: 100px !important;
-  border-radius: 12px !important;
-  margin: var(--margin-small);
-}
-
-.back-button:hover {
-  color: var(--color-error) !important;
-}
-
-.create-button:hover {
-  color: var(--color-success) !important;
-
-}
