+++ index.html
@@ -17,8 +17,9 @@
           rel="stylesheet"
   />
   <!-- css files -->
-  <link href="global.css" rel="stylesheet" />
-  <link href="index.css" rel="stylesheet" />
+  <link rel="stylesheet" href="styles/index.css">
+
+
 </head>
 <body>
 <main class="container-main">
