+++ modules/utils/domElements.js
@@ -1,0 +1,2 @@
+export const bodyElement = document.body;
+export const render = document.getElementById("render");
