+++ modules/views/renderLogin.js
@@ -1,0 +1,2 @@
+// -----login state
+export function renderLogin() {}
