+++ styles/index.css
@@ -1,0 +1,9 @@
+
+/* Base styles */
+@import "global.css";
+
+/* Components */
+@import "titleBar.css";
+@import "landingPage.css";
+@import "loginPage.css";
+@import "createAccountPage.css";
